leave unused test_instrument and test_instruments deprecated
