+++ src/qcodes/extensions/__init__.py
@@ -2,7 +2,7 @@
 The extensions module contains smaller modules that extend the functionality of QCoDeS.
 These modules may import from all of QCoDeS but do not themselves get imported into QCoDeS.
 """
-from ._driver_test_case import DriverTestCase, test_instrument, test_instruments
+from ._driver_test_case import DriverTestCase
 from ._log_export_info import log_dataset_export_info
 from .installation import register_station_schema_with_vscode
 
@@ -10,6 +10,4 @@ __all__ = [
     "register_station_schema_with_vscode",
     "log_dataset_export_info",
     "DriverTestCase",
-    "test_instrument",
-    "test_instruments",
 ]
